Use bridge networks throughout
We were using host networking for most of the containers, to support deploying on localhost. But this use case has further complications (like having to disable SSL verification) and there's no real need for it - if you want to run a service locally, run it directly. This project is intended for deploying services onto a server.